PlaneSAM/requirements.txt
absl-py==2.1.0
addict==2.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
cloudpickle==3.0.0
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.11
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
fvcore==0.1.5.post20221221
grpcio==1.65.1
huggingface-hub==0.24.5
idna==3.7
imageio==2.34.2
importlib_metadata==8.2.0
importlib_resources==6.4.0
iopath==0.1.10
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.1
MultiScaleDeformableAttention==1.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
portalocker==3.1.1
protobuf==4.25.4
pycocotools==2.0.8
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.13.1
six==1.16.0
submitit==1.5.1
tabulate==0.9.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.7.24
timm==1.0.8
tomli==2.0.1
torch==1.12.1+cu116
torchvision==0.13.1+cu116
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
yacs==0.1.8
yapf==0.40.2
zipp==3.19.2
